dependency injection reordering

+++ Assets/Code/Data/LevelData.cs
@@ -7,12 +7,12 @@ namespace SpaceEscape
     {
         [SerializeField] private string _levelName = "Level_1";
         [SerializeField, Range(0, 10)] private float _asteroidSpeed = 3.0f;
-        [SerializeField, Range(5, 100)] private float _asteroidCount = 20.0f;   // total asteroids count in the level
-        [SerializeField, Range(1, 20)] private float _asteroidDensity = 5.0f;  // number of asteroids on the screen
+        [SerializeField, Range(5, 100)] private int _asteroidCount = 20;   // total asteroids count in the level
+        [SerializeField, Range(1, 20)] private int _asteroidDensity = 5;  // number of asteroids on the screen
         
         public string LevelName => _levelName;
         public float AsteroidSpeed => _asteroidSpeed;
-        public float AsteroidCount => _asteroidCount;
-        public float AsteroidDensity => _asteroidDensity;
+        public int AsteroidCount => _asteroidCount;
+        public int AsteroidDensity => _asteroidDensity;
     }
 }
